ProgressDialog controller and fxml updated with shellfire Image
