+++ ShellfireVPN/src/main/resources/fxml/ProgressDialog.fxml
@@ -22,7 +22,7 @@
          <children>
             <Label fx:id="dynamicLabel" layoutX="7.0" layoutY="40.0" text="&lt;dynamic&gt;" AnchorPane.leftAnchor="7.0" />
             <Button fx:id="leftButton" layoutX="8.0" layoutY="124.0" minHeight="0.0" minWidth="0.0" mnemonicParsing="false" onAction="#handleLeftButton" prefHeight="25.0" prefWidth="99.0" text="Button" AnchorPane.bottomAnchor="8.0" AnchorPane.leftAnchor="8.0" />
-            <Button fx:id="rightButton" layoutX="153.0" layoutY="124.0" minHeight="0.0" minWidth="0.0" mnemonicParsing="false" onAction="#handleRightButton" prefHeight="25.0" prefWidth="99.0" text="Button" AnchorPane.bottomAnchor="8.0" AnchorPane.rightAnchor="14.0" />
+            <Button fx:id="rightButton" layoutX="153.0" layoutY="124.0" minHeight="0.0" minWidth="0.0" mnemonicParsing="false" onAction="#handleRightButton" prefHeight="25.0" prefWidth="99.0" text="Button" visible="false" AnchorPane.bottomAnchor="8.0" AnchorPane.rightAnchor="14.0" />
             <ProgressBar fx:id="progressBar" layoutY="65.0" prefHeight="18.0" prefWidth="266.0" progress="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" />
             <Label fx:id="additionTextLabel" layoutX="7.0" layoutY="14.0" minHeight="0.0" minWidth="0.0" text="heading" AnchorPane.leftAnchor="7.0" />
             <Label fx:id="bottomLabel" layoutX="7.0" layoutY="90.0" minHeight="0.0" minWidth="0.0" text="bottomLabel" AnchorPane.leftAnchor="7.0" />
