+++ ShellfireVPN/src/main/resources/fxml/ProgressDialogV.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.layout.*?>
 <?import javafx.scene.layout.AnchorPane?>
 
-<AnchorPane prefHeight="230.0" prefWidth="341.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml" fx:controller="de.shellfire.vpn.gui.controller.ProgressDiagogVController">
+<AnchorPane prefHeight="230.0" prefWidth="341.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml" fx:controller="de.shellfire.vpn.gui.controller.ProgressDialogVController">
    <children>
       <Pane fx:id="headerPane" layoutY="-3.0" prefHeight="67.0" prefWidth="305.0" AnchorPane.topAnchor="-3.0">
          <children>
